Squash #2137 - Move send notification status call to VAPROFILE to own module
