+++ app/celery/process_ses_receipts_tasks.py
@@ -1,5 +1,6 @@
 from datetime import datetime, timedelta
 
+from app.celery.send_va_profile_notification_status import check_and_queue_va_profile_email_status_callback
 import iso8601
 from app.celery.common import log_notification_total_time
 from celery.exceptions import Retry
@@ -15,12 +16,10 @@ from sqlalchemy.orm.exc import NoResultFound
 import enum
 import requests
 
-from app import notify_celery, statsd_client, va_profile_client
-from app.celery.exceptions import AutoRetryException
+from app import notify_celery, statsd_client
 from app.celery.service_callback_tasks import publish_complaint
 from app.config import QueueNames
 from app.constants import (
-    DATETIME_FORMAT,
     EMAIL_TYPE,
     HTTP_TIMEOUT,
     KEY_TYPE_NORMAL,
@@ -32,8 +31,6 @@ from app.constants import (
 )
 from app.clients.email.aws_ses import get_aws_responses
 from app.dao import notifications_dao, services_dao, templates_dao
-from app.feature_flags import FeatureFlag, is_feature_enabled
-from app.models import Notification
 from json import decoder
 from app.notifications import process_notifications
 from app.notifications.notifications_ses_callback import (
@@ -358,65 +355,3 @@ def process_ses_smtp_results(
         current_app.logger.exception(e)
         current_app.logger.error('Error processing SES SMTP results: %s', type(e))
         self.retry(queue=QueueNames.RETRY)
-
-
-def check_and_queue_va_profile_email_status_callback(notification: Notification) -> None:
-    """
-    This checks the feature flag is enabled. If it is, queues the celery task and collects data from the notification.
-    Otherwise, it only logs a message.
-
-    :param notification: the email notification to collect data from
-    """
-    current_app.logger.debug(
-        'Sending email status to VA Profile, checking feature flag... | notification %s', notification.id
-    )
-
-    if is_feature_enabled(FeatureFlag.VA_PROFILE_SMS_STATUS_ENABLED) or notification.notification_type == EMAIL_TYPE:
-        current_app.logger.debug(
-            'Sending email status to VA Profile, collecting data for notification %s', notification.id
-        )
-        notification_data = {
-            'id': str(notification.id),  # this is the notification id
-            'reference': notification.client_reference,
-            'to': notification.to,  # this is the recipient's contact info (email)
-            'status': notification.status,  # this will specify the delivery status of the notification
-            'status_reason': notification.status_reason,  # populated if there's additional context on the delivery status
-            'created_at': notification.created_at.strftime(DATETIME_FORMAT),  # noqa: F821
-            'completed_at': notification.updated_at.strftime(DATETIME_FORMAT) if notification.updated_at else None,
-            'sent_at': notification.sent_at.strftime(DATETIME_FORMAT) if notification.sent_at else None,
-            'notification_type': notification.notification_type,  # this is the channel/type of notification (email)
-            'provider': notification.sent_by,  # email provider
-        }
-
-        # data passed to tasks must be JSON serializable
-        # TODO 2137: Put back delay
-        send_email_status_to_va_profile(notification_data)
-    else:
-        current_app.logger.debug(
-            'SMS status not sent to VA Profile, feature flag disabled | notification %s', notification.id
-        )
-
-
-@notify_celery.task(
-    throws=(AutoRetryException,),
-    autoretry_for=(AutoRetryException,),
-    max_retries=60,
-    retry_backoff=True,
-    retry_backoff_max=3600,
-)
-def send_email_status_to_va_profile(notification_data: dict) -> None:
-    """
-    This function calls the VAProfileClient method to send the information to VA Profile.
-
-    :param notification_data: the email notification data to send
-    """
-
-    try:
-        va_profile_client.send_va_profile_email_status(notification_data)
-    except requests.Timeout:
-        # logging in send_va_profile_email_status
-        raise AutoRetryException
-    except requests.RequestException:
-        # logging in send_va_profile_email_status
-        # In this case the error is being handled by not retrying this celery task
-        pass
